requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cerbos==0.11.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dataclasses-json==0.6.7
ecdsa==0.19.0
fastapi==0.115.8
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mypy-extensions==1.0.0
packaging==24.2
passlib==1.7.4
protobuf==5.29.3
protoc-gen-openapiv2==0.0.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tenacity==9.0.0
types-protobuf==5.29.1.20241207
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0